% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidecgal-package.R
\docType{package}
\name{insidecgal-package}
\alias{insidecgal}
\alias{insidecgal-package}
\title{insidecgal: Point in Polygon with 'CGAL'}
\description{
Point in Polygon with ('CGAL').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dis-organization/insidecgal}
  \item Report bugs at \url{https://github.com/dis-organization/insidecgal/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
